requirements.txt
colorlog==6.9.0
homeassistant==2025.2.4
pip>=21.3.1
ruff==0.9.7
black==25.1.0
dataclasses-json==0.6.7
aiomqtt==2.3.0
python-dotenv==1.0.1
casambi-bt==0.2.4
